+++ src/js/gunpla-calculator.class.js
@@ -241,7 +241,7 @@ class GunplaCalculator {
 
   _getDisplayText(part) {
     if (part.jl) {
-      return `${part.name} (${part.jl})`; // Pilot
+      return `${part.name} (${part.jl.join('/')})`; // Pilot
     } else if (part.part.startsWith('gear-') || part.ms.startsWith('Unassigned [')) {
       return `${part.name}`; // Gear or no assigned unit
     } else if (part.name) {
@@ -266,7 +266,7 @@ class GunplaCalculator {
       this._displayPartInfo(currTarget);
 
       // Update the icons.
-      let slotText = partInput.dataset.category ? partInput.dataset.category : partInput.dataset.jl ? partInput.dataset.jl : SlotTextMap[partInput.dataset.part];
+      let slotText = partInput.dataset.category ? partInput.dataset.category : partInput.dataset.jl ? JSON.parse(partInput.dataset.jl).join('/') : SlotTextMap[partInput.dataset.part];
       let parentRow = partInput.closest('.row');
       let slotIconWrapper = parentRow.querySelector('.part-slot-type');
       slotIconWrapper.innerHTML = `<span class="slot-icon ${this._getPartSlotClass(currTarget.dataset)}" data-rarity="${currTarget.dataset.rarity}" title="${slotText}"></span>`;
@@ -296,7 +296,8 @@ class GunplaCalculator {
         }
         partClass += dataset.category.toLowerCase().replace(/\s/g, '-');
       } else if (dataset.jl) {
-        partClass += 'job-' + dataset.jl.toLowerCase();
+        const [job] = JSON.parse(dataset.jl);
+        partClass += 'job-' + job.toLowerCase();
       } else {
         partClass += SlotIconMap[dataset.part];
       }
@@ -600,7 +601,7 @@ class GunplaCalculator {
     const selectedPilot = document.querySelector('.js-select-pilot'),
         pilotInput = this.inputs.pilot;
     if (pilotInput && pilotInput.dataset.jl) {
-      const jobLicense = this._getJobLicense(pilotInput.dataset.jl);
+      const jobLicense = this._getJobLicense(pilotInput.dataset.jl[0]);
       if (jobLicense && jobLicense.multiplier && jobLicense.parameters && Array.isArray(jobLicense.parameters)) {
         jobLicense.parameters.forEach(jobParam => {
           if (jobParam in this.parametersTotal) {